-Removed routes page -Added loading progress indicator
